requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.5
    # via
    #   llama-index-core
    #   llama-index-legacy
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==23.2.0
    # via aiohttp
bcrypt==4.1.3
    # via paramiko
beautifulsoup4==4.12.3
    # via llama-index-readers-file
blinker==1.8.2
    # via flask
cachelib==0.9.0
    # via flask-caching
certifi==2024.6.2
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   geocoder
    #   nltk
cryptography==42.0.8
    # via paramiko
dataclasses-json==0.6.7
    # via
    #   llama-index-core
    #   llama-index-legacy
decorator==5.1.1
    # via ratelim
deprecated==1.2.14
    # via
    #   llama-index-core
    #   llama-index-legacy
dirtyjson==1.0.8
    # via
    #   llama-index-core
    #   llama-index-legacy
distro==1.9.0
    # via openai
flask==3.0.3
    # via flask-caching
flask-caching==2.3.0
    # via -r requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.6.0
    # via
    #   llama-index-core
    #   llama-index-legacy
future==1.0.0
    # via
    #   geocoder
    #   usaddress
geoalchemy2==0.15.1
    # via -r requirements.in
geocoder==1.38.1
    # via usaddress-scourgify
greenlet==3.0.3
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via
    #   llama-index-core
    #   llama-index-legacy
    #   llamaindex-py-client
    #   openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
joblib==1.4.2
    # via nltk
llama-index==0.10.47
    # via -r requirements.in
llama-index-agent-openai==0.2.7
    # via
    #   llama-index
    #   llama-index-program-openai
llama-index-cli==0.1.12
    # via llama-index
llama-index-core==0.10.47
    # via
    #   llama-index
    #   llama-index-agent-openai
    #   llama-index-cli
    #   llama-index-embeddings-openai
    #   llama-index-indices-managed-llama-cloud
    #   llama-index-llms-openai
    #   llama-index-multi-modal-llms-openai
    #   llama-index-program-openai
    #   llama-index-question-gen-openai
    #   llama-index-readers-file
    #   llama-index-readers-llama-parse
    #   llama-parse
llama-index-embeddings-openai==0.1.10
    # via
    #   llama-index
    #   llama-index-cli
llama-index-indices-managed-llama-cloud==0.1.6
    # via llama-index
llama-index-legacy==0.9.48
    # via llama-index
llama-index-llms-openai==0.1.22
    # via
    #   llama-index
    #   llama-index-agent-openai
    #   llama-index-cli
    #   llama-index-multi-modal-llms-openai
    #   llama-index-program-openai
    #   llama-index-question-gen-openai
llama-index-multi-modal-llms-openai==0.1.6
    # via llama-index
llama-index-program-openai==0.1.6
    # via
    #   llama-index
    #   llama-index-question-gen-openai
llama-index-question-gen-openai==0.1.3
    # via llama-index
llama-index-readers-file==0.1.25
    # via llama-index
llama-index-readers-llama-parse==0.1.4
    # via llama-index
llama-parse==0.4.4
    # via llama-index-readers-llama-parse
llamaindex-py-client==0.1.19
    # via
    #   llama-index-core
    #   llama-index-indices-managed-llama-cloud
loguru==0.7.2
    # via -r requirements.in
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
marshmallow==3.21.3
    # via dataclasses-json
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
nest-asyncio==1.6.0
    # via
    #   llama-index-core
    #   llama-index-legacy
networkx==3.3
    # via
    #   llama-index-core
    #   llama-index-legacy
nltk==3.8.1
    # via
    #   llama-index-core
    #   llama-index-legacy
numpy==1.26.4
    # via
    #   llama-index-core
    #   llama-index-legacy
    #   pandas
openai==1.35.3
    # via
    #   llama-index-agent-openai
    #   llama-index-core
    #   llama-index-legacy
packaging==24.1
    # via
    #   geoalchemy2
    #   marshmallow
pandas==2.2.2
    # via
    #   llama-index-core
    #   llama-index-legacy
paramiko==3.4.0
    # via -r requirements.in
pillow==10.3.0
    # via llama-index-core
probableparsing==0.0.1
    # via usaddress
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.7.4
    # via
    #   llamaindex-py-client
    #   openai
pydantic-core==2.18.4
    # via pydantic
pynacl==1.5.0
    # via paramiko
pypdf==4.2.0
    # via llama-index-readers-file
python-crfsuite==0.9.10
    # via usaddress
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   realtime
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   llama-index-core
    #   yaml-config
ratelim==0.1.6
    # via geocoder
realtime==1.0.6
    # via -r requirements.in
regex==2024.5.15
    # via
    #   nltk
    #   tiktoken
requests==2.32.3
    # via
    #   geocoder
    #   llama-index-core
    #   llama-index-legacy
    #   tiktoken
sentry-sdk==2.6.0
    # via -r requirements.in
six==1.16.0
    # via
    #   geocoder
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
sqlalchemy[asyncio]==2.0.31
    # via
    #   -r requirements.in
    #   geoalchemy2
    #   llama-index-core
    #   llama-index-legacy
striprtf==0.0.26
    # via llama-index-readers-file
tenacity==8.4.1
    # via
    #   llama-index-core
    #   llama-index-legacy
tiktoken==0.7.0
    # via
    #   llama-index-core
    #   llama-index-legacy
tqdm==4.66.4
    # via
    #   llama-index-core
    #   nltk
    #   openai
typing-extensions==4.12.2
    # via
    #   llama-index-core
    #   llama-index-legacy
    #   openai
    #   pydantic
    #   pydantic-core
    #   realtime
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via
    #   dataclasses-json
    #   llama-index-core
    #   llama-index-legacy
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   requests
    #   sentry-sdk
usaddress==0.5.10
    # via usaddress-scourgify
usaddress-scourgify==0.6.0
    # via -r requirements.in
websockets==12.0
    # via realtime
werkzeug==3.0.3
    # via flask
wrapt==1.16.0
    # via
    #   deprecated
    #   llama-index-core
yaml-config==0.1.5
    # via usaddress-scourgify
yarl==1.9.4
    # via aiohttp
